Refactoring(server-info): base __init__/__rate for every subsystem
Added: disk/net subsystems
