+++ netutils_linux_hardware/subsystem.py
@@ -2,9 +2,15 @@
 
 from abc import abstractmethod
 
+from netutils_linux_hardware.folding import Folding
+
 
 class Subsystem(object):
     """ Base class for CPU, Memory, etc """
+
+    def __init__(self, data=None, folding=None):
+        self.data = data
+        self.folding = folding
 
     @abstractmethod
     def collect(self):
@@ -17,6 +23,10 @@ class Subsystem(object):
         pass
 
     @abstractmethod
-    def rate(self, rater):
+    def rate(self):
         """ Rating every detail in that data """
         pass
+
+    def __rate(self, func, key):
+        items = self.data.get(key)
+        return self.folding.fold(dict((item, func(item)) for item in items), Folding.SUBSYSTEM) if items else 1
